fixed utilization function in processor.cpp

+++ include/processor.h
@@ -4,8 +4,8 @@
 class Processor {
  public:
   float Utilization();  // TODO: See src/processor.cpp
-  long user, nice, system, idle, iowait, irq, softirq, steal, guest, guest_nice;
-  long prevuser, prevnice, prevsystem, previdle, previowait, previrq,
+  float user, nice, system, idle, iowait, irq, softirq, steal, guest, guest_nice;
+  float prevuser, prevnice, prevsystem, previdle, previowait, previrq,
       prevsoftirq, prevsteal;
   float cpu_percent;
   // TODO: Declare any necessary private members
